+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/TransactionRequestListener.kt
@@ -9,14 +9,17 @@ package co.omisego.omisego.websocket.listener
 
 import co.omisego.omisego.model.APIError
 import co.omisego.omisego.model.transaction.consumption.TransactionConsumption
+import co.omisego.omisego.model.transaction.request.TransactionRequest
 import co.omisego.omisego.websocket.event.SocketEvent
 import co.omisego.omisego.websocket.event.TransactionConsumptionFinalizedEvent
 import co.omisego.omisego.websocket.event.TransactionConsumptionRequestEvent
 import co.omisego.omisego.websocket.strategy.FilterStrategy
 
 abstract class TransactionRequestListener(
-    final override val strategy: FilterStrategy = FilterStrategy.Event(allowedEvents)
+    transactionRequest: TransactionRequest
 ) : SimpleSocketCustomEventListener<SocketEvent<*>>() {
+    final override val strategy: FilterStrategy = FilterStrategy.Topic(transactionRequest.socketTopic)
+
     final override fun onSpecificEvent(event: SocketEvent<*>) {
         when (event) {
             is TransactionConsumptionRequestEvent -> event.socketReceive.data?.let(::onTransactionConsumptionRequest)
@@ -30,11 +33,4 @@ abstract class TransactionRequestListener(
     abstract fun onTransactionConsumptionRequest(transactionConsumption: TransactionConsumption)
     abstract fun onTransactionConsumptionFinalizedSuccess(transactionConsumption: TransactionConsumption)
     abstract fun onTransactionConsumptionFinalizedFail(transactionConsumption: TransactionConsumption, apiError: APIError)
-
-    companion object {
-        internal val allowedEvents = listOf(
-            TransactionConsumptionRequestEvent::class.java,
-            TransactionConsumptionFinalizedEvent::class.java
-        )
-    }
 }