+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -14,7 +15,7 @@ and `OmiseGO` adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 ### Changed
 - Change `TransactionExchange` rate to nullable type
 - Deprecated `setConnectionListener` and `setChannelListener` as a compilation error.
-- Deprecated `startListeningEvents` as a compilation error.
+- Deprecated `SocketCustomEventListener.TransactionRequestListener` and `SocketCustomEventListener.TransactionConsumptionListener` as a compilation error.
 
 ### Fixed
 - [Fix the websocket client doesn't rejoin the channel automatically when the network is temporarily offline.](https://github.com/omisego/android-sdk/issues/63)